requirements.txt
python-xrd